checking if refresh expired
