+++ src/python/sense/client/client.py
@@ -1,6 +1,6 @@
 from sense.client.driverclient import DriverClient
 from sense.client.discoveryclient import DiscoveryClient
 
-class Client(DriverClient, DiscoveryClient):
+class Client(DriverClient, ServiceClient, DiscoveryClient):
     def __init__(self):
         super(Client, self).__init__()
